fix: 체크리스트 아이템 문자열 변환 및 로깅 개선
- 체크리스트 아이템을 문자열로 안전하게 변환하는 로직 추가
- 로깅 시 아이템의 텍스트를 올바르게 표시하도록 수정
- 검색 결과가 부족할 경우의 로깅 메시지 개선

+++ app/services/gemini/api_client.py
@@ -16,7 +16,7 @@ import google.generativeai as genai
 
 from app.core.config import settings
 from app.prompts.enhanced_prompts import get_enhanced_knowledge_prompt
-from .config import GeminiConfig, GeminiAPIError, SearchResult
+from .config import GeminiConfig, GeminiAPIError, GeminiResponseError, SearchResult
 from .utils import create_error_result
 
 logger = logging.getLogger(__name__)
@@ -117,7 +117,7 @@ class GeminiApiClient:
                     model_with_search.generate_content,
                     prompt,
                     generation_config=genai.types.GenerationConfig(
-                        max_output_tokens=GeminiConfig.MAX_OUTPUT_TOKENS * 0.8,  # 검색 기능 사용 시 토큰 여유
+                        max_output_tokens=int(GeminiConfig.MAX_OUTPUT_TOKENS * 0.8),  # 검색 기능 사용 시 토큰 여유
                         temperature=GeminiConfig.TEMPERATURE,
                         top_p=GeminiConfig.TOP_P,
                         top_k=GeminiConfig.TOP_K,
